PB-1894: Resolve cyclic import
