+++ app/routes.py
@@ -9,8 +9,8 @@ from flask import jsonify
 from flask import make_response
 from flask import request
 
-from app import app
 from app import settings
+from app.app import app
 from app.helpers.s3 import get_s3_file
 from app.helpers.wms import get_wms_backend_readiness
 from app.helpers.wmts import prepare_wmts_cached_response
